refactor(api): move agent-schedules to modules/ and add subpath import aliases
