+++ apps/api/src/modules/agent-schedules/index.ts
@@ -1,11 +1,20 @@
 import { Elysia, t } from 'elysia';
-import { authContext } from '../shared/auth-context';
-import { guards } from '../shared/guards';
-import { noContent } from '../shared/http';
-import { HttpError } from '../shared/lib';
-import { accessErrors, commonErrors } from '../shared/responses';
-import { mcpTool } from '../mcp/generate';
+import { authContext } from '#shared/auth-context';
+import { guards } from '#shared/guards';
+import { noContent } from '#shared/http';
+import { HttpError } from '#shared/lib';
+import { accessErrors, commonErrors } from '#shared/responses';
+import { mcpTool } from '#mcp/generate';
 import { nextCronRun } from './cron';
+import {
+  AgentScheduleResponse,
+  AgentScheduleListResponse,
+  QueuedRunResponse,
+  ScheduleRunListResponse,
+  createScheduleBody,
+  scheduleParams,
+  updateScheduleBody,
+} from './model';
 import {
   createAgentSchedule,
   deleteAgentSchedule,
@@ -14,65 +23,7 @@ import {
   listAgentSchedules,
   listScheduleRuns,
   updateAgentSchedule,
-} from './store';
-
-const params = t.Object({
-  projectKey: t.String(),
-  scheduleId: t.Numeric({ description: 'Schedule id from list_agent_schedules.' }),
-});
-const status = t.UnionEnum(['active', 'paused'], {
-  description: "'active' runs on the cron, 'paused' does not run until it is set back to 'active'.",
-});
-const scheduleBody = t.Object({
-  agentId: t.Number({ description: 'Internal agent id from list_ai_agents that runs the task.' }),
-  name: t.String({ minLength: 1, maxLength: 120, description: 'Display name of the schedule.' }),
-  prompt: t.String({
-    minLength: 1,
-    maxLength: 20_000,
-    description: 'Task sent to the agent on every run.',
-  }),
-  cron: t.String({
-    minLength: 1,
-    maxLength: 120,
-    description: "Five-field cron expression in UTC, e.g. '0 9 * * 1' for Mondays at 09:00.",
-  }),
-  status: t.Optional(status),
-});
-
-// A schedule DTO (AgentScheduleRow from the store).
-const AgentScheduleResponse = t.Object({
-  id: t.Number(),
-  agentId: t.Number(),
-  agentName: t.String(),
-  name: t.String(),
-  prompt: t.String(),
-  cron: t.String(),
-  timezone: t.Literal('UTC'),
-  status,
-  nextRunAt: t.String(),
-  lastRunAt: t.Nullable(t.String()),
-  lastRunStatus: t.Nullable(t.String()),
-  createdAt: t.String(),
-  updatedAt: t.String(),
-});
-
-// One run of a schedule (ScheduleRunRow from the store), with the agent's answer in
-// `output` once the run has finished.
-const ScheduleRunResponse = t.Object({
-  id: t.Number(),
-  status: t.String(),
-  trigger: t.String(),
-  prompt: t.String(),
-  attempts: t.Number(),
-  lastError: t.Nullable(t.String()),
-  output: t.Nullable(t.String()),
-  scheduledFor: t.Nullable(t.String()),
-  startedAt: t.Nullable(t.String()),
-  finishedAt: t.Nullable(t.String()),
-  createdAt: t.String(),
-});
-
-const QueuedRunResponse = t.Object({ runId: t.Number() });
+} from './service';
 
 function requiredText(value: string, field: string): string {
   const trimmed = value.trim();
@@ -88,7 +39,7 @@ export const agentScheduleRoutes = new Elysia({
   .use(guards)
   .get('/projects/:projectKey/agent-schedules', ({ project }) => listAgentSchedules(project.id), {
     permission: ['ai_agents', 'read'],
-    response: { 200: t.Array(AgentScheduleResponse), ...accessErrors },
+    response: { 200: AgentScheduleListResponse, ...accessErrors },
     detail: {
       summary: 'List agent schedules',
       description: "List the project's agent schedules with their cron, next run, and last run.",
@@ -113,7 +64,7 @@ export const agentScheduleRoutes = new Elysia({
       return row;
     },
     {
-      body: scheduleBody,
+      body: createScheduleBody,
       permission: ['ai_agents', 'create'],
       response: { 201: AgentScheduleResponse, ...commonErrors },
       detail: {
@@ -146,8 +97,8 @@ export const agentScheduleRoutes = new Elysia({
       return row;
     },
     {
-      params,
-      body: t.Partial(scheduleBody),
+      params: scheduleParams,
+      body: updateScheduleBody,
       permission: ['ai_agents', 'edit'],
       response: { 200: AgentScheduleResponse, ...commonErrors },
       detail: {
@@ -166,7 +117,7 @@ export const agentScheduleRoutes = new Elysia({
       return noContent();
     },
     {
-      params,
+      params: scheduleParams,
       permission: ['ai_agents', 'delete'],
       response: { 204: t.Void(), ...commonErrors },
       detail: {
@@ -185,7 +136,7 @@ export const agentScheduleRoutes = new Elysia({
       return { runId };
     },
     {
-      params,
+      params: scheduleParams,
       permission: ['ai_agents', 'edit'],
       response: { 202: QueuedRunResponse, ...commonErrors },
       detail: {
@@ -205,9 +156,9 @@ export const agentScheduleRoutes = new Elysia({
       return rows;
     },
     {
-      params,
+      params: scheduleParams,
       permission: ['ai_agents', 'read'],
-      response: { 200: t.Array(ScheduleRunResponse), ...commonErrors },
+      response: { 200: ScheduleRunListResponse, ...commonErrors },
       detail: {
         summary: 'List agent schedule runs',
         description:
