+++ apps/api/src/projects/copy.ts
@@ -32,8 +32,8 @@ import {
   listAgentSkills,
 } from '../agent-skills/store';
 import { listAgentToolLinks, setAgentTools } from '../agent-tools/store';
-import { listAgentSchedules, createAgentSchedule } from '../agent-schedules/store';
-import { nextCronRun } from '../agent-schedules/cron';
+import { listAgentSchedules, createAgentSchedule } from '#modules/agent-schedules/service';
+import { nextCronRun } from '#modules/agent-schedules/cron';
 import { getObject } from '../shared/s3';
 
 // Which parts of a source project the copy carries over. Each key mirrors a section
